+++ Package.swift
@@ -12,13 +12,15 @@ let package = Package(
     ],
     dependencies: [
         .package(url: "https://github.com/swiftlang/swift-docc-plugin.git", from: "1.4.0"),
-        .package(url: "https://github.com/bare-swift/swift-bytes.git", from: "0.1.0")
+        .package(url: "https://github.com/bare-swift/swift-bytes.git", from: "0.1.0"),
+        .package(url: "https://github.com/apple/swift-crypto.git", from: "3.0.0")
     ],
     targets: [
         .target(
             name: "OAuth2Client",
             dependencies: [
-                .product(name: "Bytes", package: "swift-bytes")
+                .product(name: "Bytes", package: "swift-bytes"),
+                .product(name: "Crypto", package: "swift-crypto")
             ]
         ),
         .testTarget(